# backend/openapi/openapi.json
{
  "components": {
    "schemas": {
...
      "CategoryAxisParam": {
        "enum": [
          "content-type",
          "use-case",
          "technique",
          "style",
          "subject"
        ],
        "title": "CategoryAxisParam",
        "type": "string"
      },
...
      "ProblemSchema": {
        "additionalProperties": false,
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "detail": {
            "title": "Detail",
            "type": "string"
          },
          "errors": {
            "items": {
              "$ref": "#/components/schemas/ProblemFieldErrorSchema"
            },
            "title": "Errors",
            "type": "array"
          },
          "instance": {
            "title": "Instance",
            "type": "string"
          },
          "status": {
            "title": "Status",
            "type": "integer"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "traceId": {
            "title": "Traceid",
            "type": "string"
          },
          "type": {
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "type",
          "title",
          "status",
          "code",
          "detail",
          "instance",
          "traceId",
          "errors"
        ],
        "title": "ProblemSchema",
        "type": "object"
      },
...
      "PromptSort": {
        "enum": [
          "relevance",
          "trending",
          "value",
          "newest"
        ],
        "title": "PromptSort",
        "type": "string"
      },
...
      "PublicEnvelope_CategoryProjectionSchema_": {
        "additionalProperties": false,
        "properties": {
          "data": {
            "$ref": "#/components/schemas/CategoryProjectionSchema"
          },
          "meta": {
            "$ref": "#/components/schemas/ResponseMetaSchema"
          }
        },
        "required": [
          "data",
          "meta"
        ],
        "title": "PublicEnvelope[CategoryProjectionSchema]",
        "type": "object"
      },
      "PublicEnvelope_FacetSetSchema_": {
        "additionalProperties": false,
        "properties": {
          "data": {
            "$ref": "#/components/schemas/FacetSetSchema"
          },
          "meta": {
            "$ref": "#/components/schemas/ResponseMetaSchema"
          }
        },
        "required": [
          "data",
          "meta"
        ],
        "title": "PublicEnvelope[FacetSetSchema]",
        "type": "object"
      },
...
      "Window": {
        "enum": [
          "7d",
          "30d",
          "all"
        ],
        "title": "Window",
        "type": "string"
      },
...
    }
  },
  "info": {
    "description": "Anonymous, read-only API derived from one immutable catalog revision. The beta adapter reads the verified Git publication manifest and is replaceable through the CatalogRepository port.",
    "summary": "Internal-beta Prompt catalog API",
    "title": "pSEO Public Read API",
    "version": "1.0.0-beta.1"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/v1/categories/{axis}/{slug}": {
      "get": {
        "operationId": "category_projection_api_v1_categories__axis___slug__get",
        "parameters": [
          {
            "in": "path",
            "name": "axis",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/CategoryAxisParam"
            }
          },
          {
            "in": "path",
            "name": "slug",
            "required": true,
            "schema": {
              "title": "Slug",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "locale",
            "required": true,
            "schema": {
              "maxLength": 35,
              "minLength": 2,
              "title": "Locale",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 1024,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 24,
              "maximum": 50,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "sort",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/PromptSort",
              "default": "value"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicEnvelope_CategoryProjectionSchema_"
                }
              }
            },
            "description": "Successful Response",
            "headers": {
              "ETag": {
                "schema": {
                  "type": "string"
                }
              },
              "X-Content-Revision": {
                "schema": {
                  "type": "string"
                }
              },
              "X-Request-ID": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "304": {
            "description": "Not modified",
            "headers": {
              "ETag": {
                "schema": {
                  "type": "string"
                }
              },
              "X-Content-Revision": {
                "schema": {
                  "type": "string"
                }
              },
              "X-Request-ID": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "400": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemSchema"
                }
              }
            },
            "description": "Problem details"
          },
          "404": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemSchema"
                }
              }
            },
            "description": "Problem details"
          },
          "410": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemSchema"
                }
              }
            },
            "description": "Problem details"
          },
          "422": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemSchema"
                }
              }
            },
            "description": "Validation Error"
          },
          "503": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemSchema"
                }
              }
            },
            "description": "Problem details"
          }
        },
        "summary": "Read an L2 category projection",
        "tags": [
          "projections"
        ]
      }
    },
    "/api/v1/facets": {
      "get": {
        "operationId": "facets_api_v1_facets_get",
        "parameters": [
          {
            "in": "query",
            "name": "locale",
            "required": true,
            "schema": {
              "maxLength": 35,
              "minLength": 2,
              "title": "Locale",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "q",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 200,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Q"
            }
          },
          {
            "in": "query",
            "name": "contentType",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "items": {
                    "type": "string"
                  },
                  "type": "array"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Contenttype"
            }
          },
          {
            "in": "query",
            "name": "model",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "items": {
                    "type": "string"
                  },
                  "type": "array"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Model"
            }
          },
          {
            "in": "query",
            "name": "useCase",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "items": {
                    "type": "string"
                  },
                  "type": "array"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Usecase"
            }
          },
          {
            "in": "query",
            "name": "technique",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "items": {
                    "type": "string"
                  },
                  "type": "array"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Technique"
            }
          },
          {
            "in": "query",
            "name": "style",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "items": {
                    "type": "string"
                  },
                  "type": "array"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Style"
            }
          },
          {
            "in": "query",
            "name": "subject",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "items": {
                    "type": "string"
                  },
                  "type": "array"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Subject"
            }
          },
          {
            "in": "query",
            "name": "creator",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "items": {
                    "type": "string"
                  },
                  "type": "array"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Creator"
            }
          },
          {
            "in": "query",
            "name": "window",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/Window",
              "default": "all"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicEnvelope_FacetSetSchema_"
                }
              }
            },
            "description": "Successful Response",
            "headers": {
              "ETag": {
                "schema": {
                  "type": "string"
                }
              },
              "X-Content-Revision": {
                "schema": {
                  "type": "string"
                }
              },
              "X-Request-ID": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "304": {
            "description": "Not modified",
            "headers": {
              "ETag": {
                "schema": {
                  "type": "string"
                }
              },
              "X-Content-Revision": {
                "schema": {
                  "type": "string"
                }
              },
              "X-Request-ID": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "400": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemSchema"
                }
              }
            },
            "description": "Problem details"
          },
          "422": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemSchema"
                }
              }
            },
            "description": "Validation Error"
          },
          "429": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemSchema"
                }
              }
            },
            "description": "Problem details"
          },
          "503": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemSchema"
                }
              }
            },
            "description": "Problem details"
          }
        },
        "summary": "Read facet counts for the current filtered result set",
        "tags": [
          "prompts"
        ]
      }
    },
    "/api/v1/home": {
      "get": {
        "operationId": "home_api_v1_home_get",
        "parameters": [
          {
            "in": "query",
            "name": "locale",
            "required": true,
            "schema": {
              "maxLength": 35,
              "minLength": 2,
              "title": "Locale",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicEnvelope_HomeDataSchema_"
                }
              }
            },
            "description": "Successful Response",
            "headers": {
              "ETag": {
                "schema": {
                  "type": "string"
                }
              },
              "X-Content-Revision": {
                "schema": {
                  "type": "string"
                }
              },
              "X-Request-ID": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "304": {
            "description": "Not modified",
            "headers": {
              "ETag": {
                "schema": {
                  "type": "string"
                }
              },
              "X-Content-Revision": {
                "schema": {
                  "type": "string"
                }
              },
              "X-Request-ID": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "400": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemSchema"
                }
              }
            },
            "description": "Problem details"
          },
          "404": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemSchema"
                }
              }
            },
            "description": "Problem details"
          },
          "422": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemSchema"
                }
              }
            },
            "description": "Validation Error"
          },
          "503": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemSchema"
                }
              }
            },
            "description": "Problem details"
          }
        },
        "summary": "Read the L1 Prompt hub projection",
        "tags": [
          "catalog"
        ]
      }
    },
    "/api/v1/locales": {
      "get": {
        "operationId": "locales_api_v1_locales_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicEnvelope_list_LocaleInfoSchema__"
                }
              }
            },
            "description": "Successful Response",
            "headers": {
              "ETag": {
                "schema": {
                  "type": "string"
                }
              },
              "X-Content-Revision": {
                "schema": {
                  "type": "string"
                }
              },
              "X-Request-ID": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "304": {
            "description": "Not modified",
            "headers": {
              "ETag": {
                "schema": {
                  "type": "string"
                }
              },
              "X-Content-Revision": {
                "schema": {
                  "type": "string"
                }
              },
              "X-Request-ID": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "400": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemSchema"
                }
              }
            },
            "description": "Problem details"
          },
          "503": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemSchema"
                }
              }
            },
            "description": "Problem details"
          }
        },
        "summary": "List configured locales",
        "tags": [
          "catalog"
        ]
      }
    },
    "/api/v1/models/{slug}": {
      "get": {
        "operationId": "model_projection_api_v1_models__slug__get",
        "parameters": [
          {
            "in": "path",
            "name": "slug",
            "required": true,
            "schema": {
              "title": "Slug",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "locale",
            "required": true,
            "schema": {
              "maxLength": 35,
              "minLength": 2,
              "title": "Locale",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 1024,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 24,
              "maximum": 50,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "sort",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/PromptSort",
              "default": "value"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicEnvelope_ModelProjectionSchema_"
                }
              }
            },
            "description": "Successful Response",
            "headers": {
              "ETag": {
                "schema": {
                  "type": "string"
                }
              },
              "X-Content-Revision": {
                "schema": {
                  "type": "string"
                }
              },
              "X-Request-ID": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "304": {
            "description": "Not modified",
            "headers": {
              "ETag": {
                "schema": {
                  "type": "string"
                }
              },
              "X-Content-Revision": {
                "schema": {
                  "type": "string"
                }
              },
              "X-Request-ID": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "400": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemSchema"
                }
              }
            },
            "description": "Problem details"
          },
          "404": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemSchema"
                }
              }
            },
            "description": "Problem details"
          },
          "410": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemSchema"
                }
              }
            },
            "description": "Problem details"
          },
          "422": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemSchema"
                }
              }
            },
            "description": "Validation Error"
          },
          "503": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemSchema"
                }
              }
            },
            "description": "Problem details"
          }
        },
        "summary": "Read an L3 model projection",
        "tags": [
          "projections"
        ]
      }
    },
...
  },
  "servers": [
    {
      "description": "Local beta",
      "url": "http://127.0.0.1:8000"
    }
  ]
}